requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
asyncpg==0.20.1
envparse==0.2.0
flake8==3.7.9
httptools==0.0.13
httpx==0.9.3
pyjwt==1.7.1
ujson==1.35
uvloop==0.14.0
validators==0.14.2
websockets==8.1
yandex-pgmigrate==1.0.4
